[WFLY-10711] Observability: Support MicroProfile Health
In order to improve Observability of WildFly on OpenShift, WildFly needs
to expose health check probes
so that the container can query whether WildFly and its deployed
applications are healthy or not.

This will be implemented by supporting Eclipse MicroProfile Health 1.0
for applications deployed in WildFly.
The integration uses the smallrye-health component to provide the
MicroProfile Health implementation.

JIRA: https://issues.jboss.org/browse/WFLY-10711